refactor code, check case more and addColor

+++ src/main/java/sit/int221/integratedproject/kanbanborad/services/StatusService.java
@@ -1,17 +1,13 @@
 package sit.int221.integratedproject.kanbanborad.services;
 
 import jakarta.transaction.Transactional;
-import jakarta.validation.ConstraintViolationException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import sit.int221.integratedproject.kanbanborad.dtos.request.StatusRequestDTO;
-import sit.int221.integratedproject.kanbanborad.dtos.request.StatusTransferRequestDTO;
-import sit.int221.integratedproject.kanbanborad.dtos.request.TaskRequestDTO;
+import sit.int221.integratedproject.kanbanborad.dtos.response.StatusAddEditResponseDTO;
 import sit.int221.integratedproject.kanbanborad.dtos.response.StatusResponseDTO;
 import sit.int221.integratedproject.kanbanborad.dtos.response.StatusResponseDetailDTO;
-import sit.int221.integratedproject.kanbanborad.dtos.response.TaskAddEditResponseDTO;
-import sit.int221.integratedproject.kanbanborad.dtos.response.TaskResponseDTO;
 import sit.int221.integratedproject.kanbanborad.entities.Status;
 import sit.int221.integratedproject.kanbanborad.entities.Task;
 import sit.int221.integratedproject.kanbanborad.exceptions.BadRequestException;
@@ -50,6 +46,7 @@ public class StatusService {
                     StatusResponseDetailDTO statusResponseDTO  = modelMapper.map(status, StatusResponseDetailDTO.class);
                     statusResponseDTO.setName(Utils.trimString(status.getName()));
                     statusResponseDTO.setDescription(Utils.trimString(status.getDescription()));
+                    statusResponseDTO.setColor(Utils.trimString(status.getColor()));
                     statusResponseDTO.setCountTask(status.getTasks().size());
                     return statusResponseDTO;
                 })
@@ -57,36 +54,50 @@ public class StatusService {
     }
 
     @Transactional
-    public StatusResponseDTO createNewStatus(StatusRequestDTO statusDTO) {
+    public StatusAddEditResponseDTO createNewStatus(StatusRequestDTO statusDTO) {
         Status status = new Status();
         status.setName(Utils.trimString(statusDTO.getName()));
         status.setDescription(Utils.trimString(statusDTO.getDescription()));
+        status.setColor(Utils.trimString(statusDTO.getColor()));
         List<Status> statuses = statusRepository.findAll();
         for (Status statusLoop : statuses) {
             if (statusLoop.getName().equals(status.getName())) {
                 throw new BadRequestException("Could not execute due to duplicate status name.");
             }
+            if (statusLoop.getColor().equals(status.getColor())) {
+                throw new BadRequestException("Could not execute due to duplicate color name.");
+            }
         }
 
         Status savedStatus = statusRepository.save(status);
-        return modelMapper.map(savedStatus, StatusResponseDTO.class);
+        return modelMapper.map(savedStatus, StatusAddEditResponseDTO.class);
     }
 
     @Transactional
-    public StatusResponseDTO updateStatus(Integer id, StatusRequestDTO statusDTO) {
+    public StatusAddEditResponseDTO updateStatus(Integer id, StatusRequestDTO statusDTO) {
+        if (statusDTO.getName().equals("NO_STATUS")) {
+            throw new GeneralException("can not edit NO_STATUS status");
+        }
         Status existingStatus = statusRepository.findById(id)
                         .orElseThrow(() -> new ItemNotFoundException("Status Id " + id + " DOES NOT EXIST !!!"));
         existingStatus.setName(Utils.trimString(statusDTO.getName()));
         existingStatus.setDescription(Utils.trimString(statusDTO.getDescription()));
+        existingStatus.setColor(Utils.trimString(statusDTO.getColor()));
+
         Status updatedStatus = statusRepository.save(existingStatus);
-        return modelMapper.map(updatedStatus, StatusResponseDTO.class);
+        return modelMapper.map(updatedStatus, StatusAddEditResponseDTO.class);
     }
 
     @Transactional
     public StatusResponseDTO deleteStatus(Integer id) {
         Status statusToDelete = statusRepository.findById(id)
                 .orElseThrow(() -> new ItemNotFoundException("Status Id " + id + " DOES NOT EXIST !!!"));
-
+        if (!statusToDelete.getTasks().isEmpty()) {
+            throw new GeneralException("can not delete cuz status has task");
+        }
+        if (statusToDelete.getName().equals("NO_STATUS")) {
+            throw new GeneralException("can not delete NO_STATUS status");
+        }
         statusRepository.deleteById(id);
         return modelMapper.map(statusToDelete, StatusResponseDTO.class);
     }
@@ -102,6 +113,9 @@ public class StatusService {
         for (Task task : tasks) {
             task.setStatus(transferStatus);
             taskRepository.save(task);
+        }
+        if (statusToDelete.getName().equals("NO_STATUS")) {
+            throw new GeneralException("can not delete NO_STATUS status");
         }
 
         statusRepository.deleteById(statusToDelete.getId());
